Add Report model and plug keyword inside it
At the top level of the data model we have reports.

You create a report by passing multiple keywords, one by line.  A report
thus has many keywords. Note that multiple keyword objects with the same
keyword string may exist in the database; we do this to prevent a later
report to modify a former report that happens to have the same keyword.

When a report is created, the related keywords are created,
and when a keyword is created the related urls are created.
When everything has been fetched it is committed one-shot as
a big transaction inside the database.

The Report index is a standard rails' index view. The Report show
is actually a copy-paste of a Keyword index, with only the relevant
keywords being considered.